requirements.txt
alabaster==0.7.16
anyio==4.3.0
arrow==1.3.0
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bibtexparser==1.4.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
docutils==0.20.1
fake-useragent==1.5.0
Flask==3.0.2
Flask-Cors==4.0.0
free-proxy==1.1.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
imagesize==1.4.1
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==5.1.0
MarkupSafe==2.1.5
outcome==1.3.0.post0
packaging==24.0
Pygments==2.17.2
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
scholarly==1.7.11
selenium==4.18.1
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
Sphinx==7.2.6
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
trio==0.24.0
trio-websocket==0.11.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
